requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/dev.in
#
alabaster==0.7.12         # via sphinx
aniso8601==7.0.0          # via graphene
appdirs==1.4.4            # via virtualenv
astroid==2.4.2            # via pylint
babel==2.9.0              # via sphinx
certifi==2020.12.5        # via requests
cfgv==3.2.0               # via pre-commit
chardet==3.0.4            # via requests
click==7.1.2              # via flask, pip-tools
colorama==0.4.4           # via pylint, sphinx
coverage==5.3             # via -r requirements\test.in
distlib==0.3.1            # via virtualenv
docutils==0.16            # via sphinx
filelock==3.0.12          # via virtualenv
flask-graphql==2.0.1      # via -r requirements\deploy.in
flask==1.1.2              # via flask-graphql
graphene-sqlalchemy==2.3.0  # via -r requirements\deploy.in
graphene==2.1.8           # via -r requirements\deploy.in, graphene-sqlalchemy
graphql-core==2.3.2       # via flask-graphql, graphene, graphql-relay, graphql-server-core
graphql-relay==2.0.1      # via graphene
graphql-server-core==1.2.0  # via flask-graphql
identify==1.5.10          # via pre-commit
idna==2.10                # via requests
imagesize==1.2.0          # via sphinx
importlib-metadata==3.1.1  # via pre-commit, virtualenv
isort==5.6.4              # via pylint
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask, sphinx
lazy-object-proxy==1.4.3  # via astroid
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via pylint
nodeenv==1.5.0            # via pre-commit
packaging==20.8           # via sphinx
pip-tools==5.4.0          # via -r requirements/dev.in
pre-commit==2.9.3         # via -r requirements/dev.in
promise==2.3              # via graphene-sqlalchemy, graphql-core, graphql-relay, graphql-server-core
psycopg2==2.8.6           # via -r requirements\deploy.in
pygments==2.7.3           # via sphinx
pylint==2.6.0             # via -r requirements\test.in
pyparsing==2.4.7          # via packaging
python-dotenv==0.15.0     # via -r requirements\deploy.in
pytz==2020.4              # via babel
pyyaml==5.3.1             # via pre-commit
requests==2.25.0          # via sphinx
rx==1.6.1                 # via graphql-core
singledispatch==3.4.0.3   # via graphene-sqlalchemy
six==1.15.0               # via astroid, graphene, graphene-sqlalchemy, graphql-core, graphql-relay, pip-tools, promise, singledispatch, virtualenv
snowballstemmer==2.0.0    # via sphinx
sphinx-rtd-theme==0.5.0   # via -r requirements\docs.in
sphinx==3.3.1             # via -r requirements\docs.in, sphinx-rtd-theme
sphinxcontrib-applehelp==1.0.2  # via sphinx
sphinxcontrib-devhelp==1.0.2  # via sphinx
sphinxcontrib-htmlhelp==1.0.3  # via sphinx
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-qthelp==1.0.3  # via sphinx
sphinxcontrib-serializinghtml==1.1.4  # via sphinx
sqlalchemy==1.3.22        # via -r requirements\deploy.in, graphene-sqlalchemy
toml==0.10.2              # via pre-commit, pylint
typed-ast==1.4.1          # via astroid
urllib3==1.26.2           # via requests
virtualenv==20.2.2        # via pre-commit
werkzeug==1.0.1           # via flask
wrapt==1.12.1             # via astroid
zipp==3.4.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
